Zombies now walk towards the player and both zombies and the player now navigates with an angle of direction which allows for more accurate movement for the zombies and possibly steering with mouse control in the future for the player.

+++ desktop/src/com/mygdx/game/desktop/Weapon.java
@@ -11,7 +11,7 @@ public class Weapon {
         this.projectileList = new ArrayList<>();
     }
 
-    public void fireWeapon(float angle, float x, float y){
+    public void fireWeapon(int angle, float x, float y){
         Projectile p = new Projectile(1000, angle, x, y);
         projectileList.add(p);
     }
